Feature/IterativeClosestPointFilter (#70)
* Added IterativeClosestPoint filter
* Added corresponding unit test
* Added util/nanoflann.hpp from DREAM3DReview
* Fixed unit test errors for InterpolatePointCloudToRegularGrid filter
* Fixed unit test errors for MapPointCloudToRegularGrid filter
* Fixed fmt::format build errors after updating vcpkg

+++ src/Plugins/ComplexCore/src/ComplexCore/Filters/ImportTextFilter.cpp
@@ -139,7 +139,7 @@ Result<> ImportTextFilter::executeImpl(DataStructure& data, const Arguments& arg
     return CsvParser::ReadFile<double, double>(inputFilePath, *dataArray, skipLines, delimiter);
   }
   default:
-    return MakeErrorResult(-1001, fmt::format("ImportTextFilter: Parameter NumericType which has a value of '{}' does not match any in complex.", numericType));
+    return MakeErrorResult(-1001, fmt::format("ImportTextFilter: Parameter NumericType which has a value of '{}' does not match any in complex.", static_cast<int8>(numericType)));
   }
 }
 } // namespace complex
